% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_module}
\alias{add_module}
\title{Add biological feature to the module}
\usage{
add_module(compendium, module, new_genes)
}
\arguments{
\item{compendium}{the compendium selected for the analysis}

\item{module}{a Module object}

\item{new_genes}{a list of gene_names}
}
\value{
a Module object
}
\description{
Add biological feature to the module
}
