% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_module}
\alias{create_module}
\title{Create a new module}
\usage{
create_module(
  compendium,
  biofeatures = NULL,
  samplesets = NULL,
  normalization = "legacy"
)
}
\arguments{
\item{compendium}{the compendium selected for the analysis}

\item{biofeatures}{list of BiologicalFeature objects from get_bf}

\item{samplesets}{a list of <samplesets> objects from get_ss}

\item{normalization}{the normalization to be used for the inference ('legacy' as default)}
}
\value{
a Module object
}
\description{
Create a new module
}
\examples{
\dontrun{
gene_names <-c('VIT_05s0094g00350','VIT_07s0031g02630','VIT_19s0015g02480','VIT_08s0007g08840')
genes <- get_bf(compendium = get_compendium(), gene_names = as.list(gene_names))
mod1 <- create_module(compendium = get_compendium, biofeatures = genes)
}
}
