% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_module_ss}
\alias{create_module_ss}
\title{Create a new module providing Sample sets}
\usage{
create_module_ss(compendium, samplesets, normalization = "legacy")
}
\arguments{
\item{compendium}{the compendium selected for the analysis}

\item{samplesets}{a list of <samplesets> objects from get_ss}

\item{normalization}{the normalization to be used for the inference ('legacy' as default)}
}
\value{
a Module object
}
\description{
Create a new module providing Sample sets
}
