% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_connection}
\alias{get_connection}
\title{Connect to a COMPASS GraphQL endpoint}
\usage{
get_connection(url = "http://compass.fmach.it/graphql")
}
\arguments{
\item{url}{a string: the current url for compass graphql API}
}
\value{
a <pycompass.compendium.Compendium> object
}
\description{
Connect to a COMPASS GraphQL endpoint
}
\examples{
conn=get_connection(url = 'http://compass.fmach.it/graphql')
}
