% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data_sources}
\alias{get_data_sources}
\title{get_data_sources}
\usage{
get_data_sources(conn = get_connection())
}
\arguments{
\item{conn}{a <pycompass.compendium.Compendium> object}
}
\value{
a data.frame
}
\description{
get_data_sources
}
\examples{
my_conn <- get_connection()
get_data_sources(conn = my_conn)
}
