% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_platform_types}
\alias{get_platform_types}
\title{get_platform_types}
\usage{
get_platform_types(conn = get_connection())
}
\arguments{
\item{conn}{a <pycompass.compendium.Compendium> object}
}
\value{
a data.frame
}
\description{
get_platform_types
}
\examples{
my_conn <- get_connection()
get_platform_types(conn = my_conn)
}
