% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ss}
\alias{get_ss}
\title{Get sample set}
\usage{
get_ss(compendium, sample_names)
}
\arguments{
\item{compendium}{the compendium selected for the analysis}

\item{sample_names}{a vector of samples ids}
}
\value{
list of SampleSets objects
}
\description{
Get sample set
}
\examples{
\dontrun{
set_01=c("GSM786264.ch1-vs-GSM786258.ch1","E-MTAB-1514.US_03_3.ch1-vs-E-MTAB-1514.US_03_1.ch1")
get_ss(compendium = get_compendium(), sample_names = set_01)
}
}
