% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_modules}
\alias{join_modules}
\title{Create a new module joining two modules}
\usage{
join_modules(module_1, module_2)
}
\arguments{
\item{module_1}{a Module object}

\item{module_2}{a Module object}
}
\value{
a Module object
}
\description{
Create a new module joining two modules
}
