% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_distribution}
\alias{plot_distribution}
\title{plot biological_features_standard_deviation_distribution}
\usage{
plot_distribution(module)
}
\arguments{
\item{module}{a "pycompass.module.Module"}
}
\value{
a "xml_document"
}
\description{
plot biological_features_standard_deviation_distribution
}
\examples{
\dontrun{
mod1 <- create_module(get_connection()$vv_compendium, genes)
my_plot_dist <- plot_distribution(mod1)
}
}
