% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Get the HTML or JSON code that plot module heatmaps}
\usage{
plot_heatmap(module)
}
\arguments{
\item{module}{a Module object}
}
\value{
a "xml_document"
}
\description{
Get the HTML or JSON code that plot module heatmaps
}
\examples{
\dontrun{
gene_names <-c('VIT_05s0094g00350','VIT_07s0031g02630','VIT_19s0015g02480','VIT_08s0007g08840')
genes <- get_bf(compendium = get_compendium(), gene_names = as.list(gene_names))
mod1 <- create_module(compendium = get_compendium, biofeatures = genes)
my_html <- plot_heatmap(mod1)
}
}
