% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_network}
\alias{plot_network}
\title{Get the HTML or JSON code that plot the module networks}
\usage{
plot_network(module, output_format = "html")
}
\arguments{
\item{module}{a Module object}

\item{output_format}{a string: either 'html' (default) or 'json'}
}
\value{
a "xml_document"
}
\description{
Get the HTML or JSON code that plot the module networks
}
\examples{
\dontrun{
gene_names <-c('VIT_05s0094g00350','VIT_07s0031g02630','VIT_19s0015g02480','VIT_08s0007g08840')
genes <- get_bf(compendium = get_compendium(), gene_names = as.list(gene_names))
mod1 <- create_module(compendium = get_compendium, biofeatures = genes)
my_plot_net <- plot_network(mod1)
}
}
