% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_module}
\alias{add_module}
\title{Title}
\usage{
add_module(compendium, module, new_genes)
}
\arguments{
\item{compendium}{the compendium selected for the analysis}

\item{module}{a "pycompass.module.Module"}

\item{new_genes}{a list of gene_names}
}
\value{
a "pycompass.module.Module"
}
\description{
Title
}
