% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_module}
\alias{create_module}
\title{create_module}
\usage{
create_module(compendium, biofeatures)
}
\arguments{
\item{compendium}{the compendium selected for the analysis}

\item{biofeatures}{a list of <pycompass.biological_feature.BiologicalFeature> objects from get_bf}
}
\value{
a "pycompass.module.Module"
}
\description{
create_module
}
\examples{
\dontrun{
mod1 <- create_module(compendium = get_connection()$vv_compendium, biofeatures = genes)
}
}
