% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bf}
\alias{get_bf}
\title{get_bf}
\usage{
get_bf(compendium, gene_names)
}
\arguments{
\item{compendium}{the compendium selected for the analysis}

\item{gene_names}{a vector of gene_ids}
}
\value{
a list of <pycompass.biological_feature.BiologicalFeature> objects
}
\description{
get_bf
}
\examples{
\dontrun{
my_conn <- get_connection(url = 'http://compass.fmach.it/graphql')
vv_compendium <- my_conn$connection$get_compendium('vitis_vinifera')
gene_names <-c('VIT_05s0094g00350','VIT_07s0031g02630','VIT_19s0015g02480','VIT_08s0007g08840')
genes <- get_bf(compendium = vv_compendium, gene_names = as.list(gene_names))
}
}
