% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_compendium}
\alias{get_compendium}
\title{get_compendium}
\usage{
get_compendium(species = "vitis_vinifera")
}
\arguments{
\item{species}{a string ('vitis_vinifera')}
}
\value{
a <pycompass.compendium.Compendium> object
}
\description{
Get a compendium by a given name, NULL otherwise
}
