% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_platform_types}
\alias{get_platform_types}
\title{Get the platform types}
\usage{
get_platform_types(compendium = get_compendium())
}
\arguments{
\item{compendium}{the compendium selected for the analysis}
}
\value{
a data.frame
}
\description{
Get the platform types
}
\examples{
vv_compendium <- get_compendium()
get_platform_types(compendium = vv_compendium)
}
