% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Get the HTML or JSON code that plot module distributions}
\usage{
plot_distribution(
  module,
  plot_type = "biological_features_standard_deviation_distribution"
)
}
\arguments{
\item{module}{a Module object}

\item{plot_type}{type of plot, select from:
- 'biological_features_standard_deviation_distribution' (default)
- 'sample_sets_magnitude_distribution'}
}
\value{
a "xml_document"
}
\description{
Get the HTML or JSON code that plot module distributions
}
\examples{
\dontrun{
gene_names <-c('VIT_05s0094g00350','VIT_07s0031g02630','VIT_19s0015g02480','VIT_08s0007g08840')
genes <- get_bf(compendium = get_compendium(), gene_names = as.list(gene_names))
mod1 <- create_module(compendium = get_compendium, biofeatures = genes)
my_plot_dist <- plot_distribution(mod1)
}
}
