% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(module)
}
\arguments{
\item{module}{a "pycompass.module.Module"}
}
\value{
a "xml_document"
}
\description{
plot_heatmap
}
\examples{
\dontrun{
mod1 <- create_module(get_connection()$vv_compendium, genes)
my_html <- plot_heatmap(mod1)
}
}
