% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rank_genes}
\alias{rank_genes}
\title{Rank all biological features on the module’s sample set using rank_method}
\usage{
rank_genes(compendium = get_compendium(), module, rank_method = "std")
}
\arguments{
\item{compendium}{the compendium selected for the analysis}

\item{module}{a "pycompass.module.Module"}

\item{rank_method}{ranking method (default 'std')}
}
\value{
a data.frame
}
\description{
Rank all biological features on the module’s sample set using rank_method
}
\examples{
 \dontrun{
ranked_genes <- rank_genes()
}
}
